remove efile default namespace
The xmlns="http://www.irs.gov/efile” attribute on the <Return> node of
most IRS e-file documents creates a problem with applying the
stylesheet transformations. Because of this namespace, the XPath
expressions within the stylesheets cannot work. Additionally, once the
XML has been parsed by JavaScript, the namespace cannot be removed. So
instead we need to remove the namespace declaration from the text
content, and we need to readjust our use of XPath for building out the
template file.