correct header declaration
The filename of the e-files is not the DLN, it’s the ObjectId. The DLN
isn’t available without accessing the AWS index properties, so we’ll
just exclude those values entirely.

But we can still use the unique identifier in the header by abusing the
other fields that display up there.