Fix XPath statements to work with IE Polyfill
XPath in IE11 is specifically a mess. It used to work with a custom
ActiveXObject that had .selectSingleNode() and .selectNodes(). However,
in IE11 that’s no longer the type of XML document returned from the
API. Instead, we can use Google’s Wicked-Good-XPath polyfill. However,
that library seems to be having DOM4 issues with descendent operators
and attributes.

To make everything work, we can manipulate the XPath statements to
generate something that will work natively and with gwxpath.