fix stylesheet processing for IE11
IE11 can use the standard DOM object for most XML operations, except stylesheet transformations. For that, we need to use a MSXML object for both the input XML and the stylesheet.

This approach attempts to minimize the complexity of using a different DOM API by serializing and re-parsing a standard DOM object at the very end. This allows us to continue to use the standard DOM API for all browsers. The XSL document has to be handled differently so that relative links throughout the document continue to work.