normalize all /mef files with 644 permissions
This updates all existing files in the /mef directory with `644`
permissions. This is necessary to automate the process of importing
files from an IRS stylesheet archive for a tax year. Many source files
(images in particular) are read-only. These cause an issue with the
commands to copy and set the attributes of the destination file.