log all stylesheet fixes applied to scripts, styles, and stylesheets
This list should be manually updated each time a new fix is applied on top
of the default IRS stylesheet files. It can serve as a reference when adding
new tax years to ensure that we're A) not overwriting changes that were
previously necessary; and B) identifying standard changes that must be
applied to new tax years.